+++ c2mon-client/c2mon-client-core/src/test/java/cern/c2mon/client/core/elasticsearch/ElasticsearchServiceTest.java
@@ -1,35 +1,20 @@
+/******************************************************************************
+ * Copyright (C) 2010-2019 CERN. All rights not expressly granted are reserved.
+ *
+ * This file is part of the CERN Control and Monitoring Platform 'C2MON'.
+ * C2MON is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation, either version 3 of the license.
+ *
+ * C2MON is distributed in the hope that it will be useful, but WITHOUT ANY
+ * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with C2MON. If not, see <http://www.gnu.org/licenses/>.
+ *****************************************************************************/
 package cern.c2mon.client.core.elasticsearch;
-
-import cern.c2mon.client.core.config.C2monClientProperties;
-import cern.c2mon.server.cache.EquipmentCache;
-import cern.c2mon.server.cache.ProcessCache;
-import cern.c2mon.server.cache.SubEquipmentCache;
-import cern.c2mon.server.cache.TagFacadeGateway;
-import cern.c2mon.server.cache.common.TagFacadeGatewayImpl;
-import cern.c2mon.server.common.datatag.DataTagCacheObject;
-import cern.c2mon.server.elasticsearch.Indices;
-import cern.c2mon.server.elasticsearch.MappingFactory;
-import cern.c2mon.server.elasticsearch.client.ElasticsearchClient;
-import cern.c2mon.server.elasticsearch.client.ElasticsearchClientImpl;
-import cern.c2mon.server.elasticsearch.config.ElasticsearchProperties;
-import cern.c2mon.server.elasticsearch.tag.config.TagConfigDocumentConverter;
-import cern.c2mon.server.elasticsearch.tag.config.TagConfigDocumentIndexer;
-import cern.c2mon.server.elasticsearch.tag.config.TagConfigDocumentListener;
-import cern.c2mon.shared.client.configuration.ConfigConstants;
-import org.apache.http.annotation.NotThreadSafe;
-import org.awaitility.Awaitility;
-import org.easymock.EasyMock;
-import org.easymock.Mock;
-import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
-import org.elasticsearch.action.admin.indices.flush.FlushRequest;
-import org.elasticsearch.node.NodeValidationException;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.Test;
-import org.powermock.reflect.Whitebox;
-import org.powermock.reflect.internal.WhiteboxImpl;
-import org.springframework.util.FileSystemUtils;
 
 import java.util.Collection;
 import java.util.Collections;
@@ -38,13 +23,40 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
+import lombok.extern.slf4j.Slf4j;
+import org.apache.http.annotation.NotThreadSafe;
+import org.awaitility.Awaitility;
+import org.easymock.Mock;
+import org.elasticsearch.action.admin.indices.delete.DeleteIndexRequest;
+import org.elasticsearch.action.admin.indices.flush.FlushRequest;
+import org.elasticsearch.node.NodeValidationException;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.springframework.util.FileSystemUtils;
+
+import cern.c2mon.client.core.config.C2monClientProperties;
+import cern.c2mon.server.cache.EquipmentCache;
+import cern.c2mon.server.cache.ProcessCache;
+import cern.c2mon.server.cache.SubEquipmentCache;
+import cern.c2mon.server.cache.TagFacadeGateway;
+import cern.c2mon.server.common.datatag.DataTagCacheObject;
+import cern.c2mon.server.elasticsearch.Indices;
+import cern.c2mon.server.elasticsearch.MappingFactory;
+import cern.c2mon.server.elasticsearch.client.ElasticsearchClientImpl;
+import cern.c2mon.server.elasticsearch.config.ElasticsearchProperties;
+import cern.c2mon.server.elasticsearch.tag.config.TagConfigDocumentConverter;
+import cern.c2mon.server.elasticsearch.tag.config.TagConfigDocumentIndexer;
+import cern.c2mon.server.elasticsearch.tag.config.TagConfigDocumentListener;
+import cern.c2mon.shared.client.configuration.ConfigConstants;
+
 import static org.easymock.EasyMock.*;
-import static org.easymock.EasyMock.createNiceMock;
-import static org.easymock.EasyMock.replay;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 @NotThreadSafe
+@Slf4j
 public class ElasticsearchServiceTest {
 
   private ElasticsearchClientImpl client;
@@ -67,23 +79,31 @@ public class ElasticsearchServiceTest {
   }
 
   @BeforeClass
-  @AfterClass
   public static void cleanup() {
+    log.debug("Clean ES properties");
     FileSystemUtils.deleteRecursively(new java.io.File(elasticsearchProperties.getEmbeddedStoragePath()));
   }
 
   @Before
   public void setupElasticsearch() throws InterruptedException, NodeValidationException {
     try {
+      log.debug("Setup Elasticsearch cluster for testing...");
       CompletableFuture<Void> nodeReady = CompletableFuture.runAsync(() -> {
         client.waitForYellowStatus();
         client.getClient().admin().indices().delete(new DeleteIndexRequest(elasticsearchProperties.getTagConfigIndex()));
         Awaitility.await().until(() -> Indices.create(elasticsearchProperties.getTagConfigIndex(), "tag_config", MappingFactory.createTagConfigMapping()));
       });
       nodeReady.get(120, TimeUnit.SECONDS);
+      log.debug("Elasticseach cluster ready for testing!");
     } catch (ExecutionException | TimeoutException e) {
       throw new RuntimeException("Timeout when waiting for embedded elasticsearch node to start!");
     }
+  }
+
+  @After
+  public void tearDown() {
+    log.debug("Closing ES client...");
+    client.close();
   }
 
   @Before
@@ -110,9 +130,9 @@ public class ElasticsearchServiceTest {
       tagDocumentListener.onConfigurationEvent(tag, ConfigConstants.Action.CREATE);
 
       client.getClient().admin().indices().flush(new FlushRequest()).actionGet();
-      
+
       ElasticsearchService service = new ElasticsearchService(properties);
-      
+
       Awaitility.await().until(() -> service.getDistinctMetadataKeys().size() > 0);
 
       assertEquals("There should be 2 tags, one for responsible and one for 1234", 2, service.getDistinctMetadataKeys().size());
@@ -155,11 +175,11 @@ public class ElasticsearchServiceTest {
       tagDocumentListener.onConfigurationEvent(tag, ConfigConstants.Action.CREATE);
 
       client.getClient().admin().indices().flush(new FlushRequest()).actionGet();
-      
+
       ElasticsearchService service = new ElasticsearchService(properties);
-      
+
       Awaitility.await().until(() -> service.findTagsByNameAndMetadata(tagname, metadataKey, testUser).size() > 0);
-      
+
       Collection<Long> tagsForResponsibleUser = service.findTagsByNameAndMetadata(tagname, metadataKey, testUser);
       assertEquals("There should be one tag with given name and metadata", 1, tagsForResponsibleUser.size());
       assertEquals(testUserTagId, tagsForResponsibleUser.stream().findFirst().get());
@@ -168,7 +188,7 @@ public class ElasticsearchServiceTest {
       throw e;
     }
   }
-  
+
   @Test
   public void testSearchByName() throws InterruptedException {
     try {