+++ c2mon-server/c2mon-server-cache/src/main/java/cern/c2mon/server/cache/alarm/oscillation/OscillationUpdateChecker.java
@@ -1,5 +1,5 @@
 /******************************************************************************
- * Copyright (C) 2010-2016 CERN. All rights not expressly granted are reserved.
+ * Copyright (C) 2010-2019 CERN. All rights not expressly granted are reserved.
  *
  * This file is part of the CERN Control and Monitoring Platform 'C2MON'.
  * C2MON is free software: you can redistribute it and/or modify it under the
@@ -17,6 +17,7 @@
 package cern.c2mon.server.cache.alarm.oscillation;
 
 import java.util.Collection;
+import java.util.Date;
 import java.util.Timer;
 import java.util.TimerTask;
 
@@ -26,17 +27,18 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.SmartLifecycle;
 import org.springframework.stereotype.Service;
 
+import lombok.extern.slf4j.Slf4j;
+
 import cern.c2mon.server.cache.AlarmCache;
 import cern.c2mon.server.cache.AliveTimerFacade;
 import cern.c2mon.server.cache.ClusterCache;
-import cern.c2mon.server.cache.DataTagCache;
-import cern.c2mon.server.cache.alarm.config.OscillationProperties;
+import cern.c2mon.server.cache.TagFacadeGateway;
 import cern.c2mon.server.cache.exception.CacheElementNotFoundException;
 import cern.c2mon.server.common.alarm.AlarmCacheObject;
 import cern.c2mon.server.common.alarm.AlarmCacheUpdater;
 import cern.c2mon.server.common.config.ServerConstants;
+import cern.c2mon.server.common.tag.Tag;
 import cern.c2mon.shared.client.alarm.AlarmQuery;
-import lombok.extern.slf4j.Slf4j;
 
 /**
  * Timer that regularly checks all the active alive timers monitoring the
@@ -47,9 +49,6 @@ import lombok.extern.slf4j.Slf4j;
  * alive-interval/3 milliseconds have expired since the last alive message
  * arrived, where alive-interval is specific to the AliveTimer object (see
  * <code>hasExpired</code> in {@link AliveTimerFacade}).
- *
- * @author Mark Brightwell
- *
  */
 @Service
 @Slf4j
@@ -57,83 +56,79 @@ public class OscillationUpdateChecker extends TimerTask implements SmartLifecycl
 
   /**
    * Cluster cache key to ensure that a server does not try to access
-   * LAST_ALIVE_TIMER_CHECK_LONG during instantiation if it already exists. This
-   * is because a cache loading blockage can happen if a server holds
-   * LAST_ALIVE_TIMER_CHECK_LONG while another server starts up. See
-   * https://issues.cern.ch/browse/TIMS-1037.
+   * LAST_CHECK_INITIALISATION_KEY during instantiation if it already exists.
    */
   private static final String LAST_CHECK_INITIALISATION_KEY = OscillationUpdateChecker.class.getName() + ".lastCheckInitialisationKey";
 
   /**
-   * Cluster cache key for retrieving the time of last check of the alives.
+   * Cluster cache key for retrieving the time of last check of the alarm oscillations.
    * Across server cluster it assures that the alive check only takes place on a
    * single server.
    */
   private static final String LAST_CHECK_LONG = OscillationUpdateChecker.class.getName() + ".lastAliveTimerCheck";
 
   /**
-   * How often the timer checks whether the alive timer have expired.
+   * How often the timer checks whether the oscillation timer have expired.
    */
-  private static final int SCAN_INTERVAL = 10000;
+  private static final long SCAN_INTERVAL = 60000L;
 
   /**
    * The time the server waits before doing first checks at start up (this gives
-   * time for incoming alives to be processed).
+   * time for incoming alarms to be processed).
    */
-  private static final int INITIAL_SCAN_DELAY = 120000;
+  private static final long INITIAL_SCAN_DELAY = 120000L;
 
   /**
    * Lifecycle flag.
    */
   private volatile boolean running = false;
 
-  /**
-   * Timer object
-   */
   private Timer timer;
 
-  /**
-   * Reference to alive timer cache.
-   */
-  private AlarmCache alarmCache;
-  
-  private DataTagCache dataTagCache;
-  
-  /**
-   * Reference to alive oscillationUpdater.
-   */
-  private OscillationUpdater oscillationUpdater;
+  private final AlarmCache alarmCache;
+
+  private final OscillationUpdater oscillationUpdater;
 
   /** Reference to the clusterCache to share values across the cluster nodes */
   private final ClusterCache clusterCache;
 
+  private final AlarmCacheUpdater alarmCacheUpdater;
+  
+  private final TagFacadeGateway tagFacadeGateway;
+  
+  private final AlarmQuery alarmCacheQuery = AlarmQuery.builder().oscillating(true).build();
+
   /**
    * Constructor.
-   * 
+   *
    * @param alarmCache
-   *          the alarm cache
+   *          the alarm cache to retrieve and update alarm cache objects.
+   * @param dataTagCache
+   *          the data tag cache to retrieve data tag objects and check their original values.
    * @param clusterCache
-   *          Reference to the clusterCache to share values accross teh cluster
-   *          nodes
+   *          the cluster cache to synchronize checks.
+   * @param oscillationUpdater
+   *          the instance that check oscillation statuses.
+   * @param AlarmCacheUpdater
+   *          the alarm cache updater.
    */
   @Autowired
-  public OscillationUpdateChecker(final AlarmCache alarmCache, final DataTagCache dataTagCache, final ClusterCache clusterCache, final OscillationUpdater oscillationUpdater) {
+  public OscillationUpdateChecker(final AlarmCache alarmCache, final ClusterCache clusterCache, final OscillationUpdater oscillationUpdater, final AlarmCacheUpdater alarmCacheUpdater, final TagFacadeGateway tagFacadeGateway) {
     super();
     this.alarmCache = alarmCache;
-    this.dataTagCache = dataTagCache;
     this.clusterCache = clusterCache;
     this.oscillationUpdater = oscillationUpdater;
+    this.alarmCacheUpdater = alarmCacheUpdater;
+    this.tagFacadeGateway = tagFacadeGateway;
   }
-
-  @Autowired
-  OscillationProperties oscillationProperties;
+  
 
   /**
-   * Initialises the clustered values
+   * Initializes the clustered values
    */
   @PostConstruct
   public void init() {
-    log.trace("Initialising AliveTimerChecker...");
+    log.trace("Initialising Alarm oscillation checker ...");
     clusterCache.acquireWriteLockOnKey(LAST_CHECK_INITIALISATION_KEY);
     try {
       if (!clusterCache.hasKey(LAST_CHECK_INITIALISATION_KEY)) {
@@ -151,7 +146,7 @@ public class OscillationUpdateChecker extends TimerTask implements SmartLifecycl
    */
   @Override
   public synchronized void start() {
-    log.info("Starting the C2MON alive timer mechanism.");
+    log.info("Starting the C2MON Alarm oscillation timer mechanism.");
     timer = new Timer("AlarmOscillationChecker");
     timer.schedule(this, INITIAL_SCAN_DELAY, SCAN_INTERVAL);
     running = true;
@@ -165,7 +160,7 @@ public class OscillationUpdateChecker extends TimerTask implements SmartLifecycl
    */
   @Override
   public synchronized void stop() {
-    log.info("Stopping the C2MON alive timer mechanism.");
+    log.info("Stopping the C2MON Alarm oscillation timer mechanism.");
     timer.cancel();
     running = false;
   }
@@ -178,46 +173,55 @@ public class OscillationUpdateChecker extends TimerTask implements SmartLifecycl
     clusterCache.acquireWriteLockOnKey(LAST_CHECK_LONG);
     try {
       Long lastCheck = (Long) clusterCache.getCopy(LAST_CHECK_LONG);
-      if (System.currentTimeMillis() - lastCheck.longValue() < 9000) { // results
-                                                                       // in
-                                                                       // check
-                                                                       // on a
-                                                                       // single
-                                                                       // server
+      if (System.currentTimeMillis() - lastCheck.longValue() < SCAN_INTERVAL - 500L) {
         log.debug("Skipping alarm oscillation check as already performed.");
       } else {
-        if (log.isDebugEnabled()) {
-          log.debug("run() : checking alarm oscillation timers ... ");
-        }
+        log.debug("checking alarm oscillation timers ... ");
         try {
-
-          AlarmQuery query = AlarmQuery.builder().oscillating(true).build();
-          Collection<Long> result = alarmCache.findAlarm(query);
-
-          for (Long alarmId : result) {
-            AlarmCacheObject alarmCopy = (AlarmCacheObject) alarmCache.getCopy(alarmId);
-            if (!oscillationUpdater.checkOscillAlive(alarmCopy)) {
-              oscillationUpdater.resetOscillationCounter(alarmCopy);
-              alarmCopy.setOscillating(false);
-              alarmCopy.setInfo(AlarmCacheUpdater.evaluateAdditionalInfo(alarmCopy, dataTagCache.get(alarmCopy.getDataTagId())));
-              alarmCache.put(alarmCopy.getId(), alarmCopy);
-            }
+          Collection<Long> oscillatingAlarmIds = alarmCache.findAlarm(alarmCacheQuery);
+          if (oscillatingAlarmIds.isEmpty()) {
+            log.debug("Currently no oscillating alarms");
+          } else {
+            log.warn("Currently {} oscillating alarms", oscillatingAlarmIds.size());
+            oscillatingAlarmIds.stream().forEach(this::updateAlarmOscillationFlag);
           }
-
-        } catch (CacheElementNotFoundException cacheEx) {
-          log.warn("Failed to locate alive timer in cache on expiration check (may happen exceptionally if just removed).", cacheEx);
         } catch (Exception e) {
-          log.error("Unexpected exception when checking the alive timers", e);
+          log.error("Unexpected exception when checking the Alarm oscillation timers", e);
         }
+
         lastCheck = Long.valueOf(System.currentTimeMillis());
         clusterCache.put(LAST_CHECK_LONG, lastCheck);
 
-        if (log.isDebugEnabled()) {
-          log.debug("run() : finished checking alarm oscillation timers ... ");
-        }
+        log.debug("finished checking alarm oscillation timers");
       } // end of else block
     } finally {
       clusterCache.releaseWriteLockOnKey(LAST_CHECK_LONG);
+    }
+  }
+
+  private void updateAlarmOscillationFlag(Long alarmId) {
+    try {
+      log.trace("Checking oscillation expiry for alarm #{}", alarmId);
+      AlarmCacheObject alarmCopy = (AlarmCacheObject) alarmCache.getCopy(alarmId);
+      if(log.isTraceEnabled()) {
+                log.trace(" -> Alarm oscillation details osc {} first osc {} count {} al ts {}", alarmCopy.isOscillating(),
+                        new Date(alarmCopy.getFirstOscTS()).toString(), alarmCopy.getCounterFault(),
+                        alarmCopy.getTimestamp().toString());
+            }
+      if (!oscillationUpdater.checkOscillAlive(alarmCopy)) {
+          log.trace(" -> ! Alarm #{} is not oscillating anymore, resetting oscillation counter", alarmId);
+          oscillationUpdater.resetOscillationCounter(alarmCopy);
+          Tag tag = tagFacadeGateway.getTag(alarmCopy.getDataTagId());
+          if(tag != null) {
+            alarmCacheUpdater.resetOscillationStatus(alarmCopy, tag);
+          } else {
+              log.error("Cannot locate data tag #{} - unable to reset oscillation status", alarmCopy.getDataTagId());
+          }
+      } else {
+          log.trace(" -> (!) Alarm #{} is still oscillating - no change", alarmId);
+      }
+    } catch (CacheElementNotFoundException e) {
+      log.error("Failed to locate corresponding tag in cache for alarm #{}. This should never happen!", alarmId);
     }
   }