+++ CHANGELOG.md
@@ -6,16 +6,45 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 All issues referenced in parentheses can be consulted under [CERN GitLab](https://gitlab.cern.ch/c2mon/c2mon/issues).
 For more details on a given release, please check also the [Milestone planning](https://gitlab.cern.ch/c2mon/c2mon/milestones?state=all).
 
+
 ## [Unreleased]
 
 ### Added
-- Client API: Added method to clear local command cache 
-- Client API: Added methods to `SupervisionService` for retrieving process and (sub-)equipment information from local cache (#238)
+- Added a warning log message that informs every minute how many alarms are currently oscillating. In case of no oscillating alarms, only a debug message is logged.
 
 ### Changed
 
 ### Fixed
 
+
+## [1.8.42] - 2019-05-28
+
+### Fixed
+- Alarm oscillation feature introduced as alpha version in [1.8.40] passed now all internal CERN stress test criterias and can be considered as stable (#145)
+- Corrected a misreferenced property name in `c2mon-server.properties` file (#261)
+
+
+## [1.8.41] - 2019-05-09
+
+### Added
+- Client API: Added method to clear local command cache (22abcb48)
+- Client API: Added methods to `SupervisionService` for retrieving process and (sub-)equipment information from local cache (#238)
+- Server, Client API: Added source timestamp to alarm object (#237)
+- Server: Added dedicated alarm logger to keep track of alarm updates in a separate log file (#229)
+
+### Changed
+- Cient API: Set ActiveMQ prefetch limit to 100 to avoid slow consumer problem (#240)
+- Updated C2MON Docker examples for Kubernetes (#205)
+
+### Fixed
+- Server: Fixed `clone()` method of `AlarmCacheObject` (5bf7a831)
+- Server: Fixed alarm state mapping to `Boolean` for database and source timestamp indexing for ElasticSearch (1627915c)
+- Server: Fixed bugs in alarm oscillation control logic introduced with last release (#233, #243, #246, #248, #145)
+- Server: Fixed problem of alarm history which logged active alarms several times with the same timestamp (#205)
+- Client API: ActiveMQ prefetch limit is now set in the code to 100 to avoid slow consumer problems (#240)
+
+### Removed
+- Removed dependency to `javax.el` validator annotations (#249)
 
 ## [1.8.40] - 2019-02-01
 **Please note**,  upgrading to this version requires changes in the database schema. This was necessary in order to implement the alarm oscillation control feature (see #145).
@@ -342,7 +371,9 @@ This patch contains bug fixes for the DAQ layer.
 - Remove of EquipmentLogger concept from DAQ Core (#56)
 
 
-[Unreleased]: https://gitlab.cern.ch/c2mon/c2mon/milestones/34
+[Unreleased]: https://gitlab.cern.ch/c2mon/c2mon/milestones/36
+[1.8.42]: https://gitlab.cern.ch/c2mon/c2mon/milestones/35
+[1.8.41]: https://gitlab.cern.ch/c2mon/c2mon/milestones/34
 [1.8.40]: https://gitlab.cern.ch/c2mon/c2mon/milestones/31
 [1.8.39]: https://gitlab.cern.ch/c2mon/c2mon/milestones/30
 [1.8.36]: https://gitlab.cern.ch/c2mon/c2mon/milestones/29